LZMR-47 implemented characters serviceinterface
