+++ Netbeans/SuperheroSightings/src/main/java/com/sg/superherosightings/controller/SightingController.java
@@ -10,6 +10,7 @@ import com.sg.superherosightings.dao.SuperheroSightingsDao;
 import com.sg.superherosightings.dto.Characters;
 import com.sg.superherosightings.dto.Location;
 import com.sg.superherosightings.dto.Sighting;
+import com.sg.superherosightings.service.CharactersService;
 import com.sg.superherosightings.service.SHService;
 import com.sg.superherosightings.service.SightingService;
 
@@ -29,13 +30,13 @@ public class SightingController {
      //Change these later to include SightingService and SightingDao
     SHService service;
     SightingService sightingService; //Change this to more specific SightingDao
-    
+    CharactersService charactersService;
     //Add Service dependency injection for Character, Location, and Organization
     
     @RequestMapping(value = {"/viewSightings"}, method = RequestMethod.GET)
     public String loadsightings(HttpServletRequest request, Model model) {
         List<Sighting> sightings = sightingService.getAllSightings();
-        List<Characters> characters = dao.getAssociatedCharacters(sightings);
+        List<Characters> characters = charactersService.getAssociatedCharacters(sightings);
         List<Location> locations = dao.getAssociatedLocations(sightings);
         model.addAttribute("sightings", sightings);
         model.addAttribute("locations", locations);
