+++ Netbeans/SuperheroSightings/src/main/java/com/sg/superherosightings/service/CharactersService.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.sg.superherosightings.dao;
+package com.sg.superherosightings.service;
 
 import com.sg.superherosightings.dto.Characters;
 import com.sg.superherosightings.dto.Sighting;
@@ -13,8 +13,8 @@ import java.util.List;
  *
  * @author laptop
  */
-public interface CharactersDao{
-    
+public interface CharactersService {
+
     public Characters addCharacter(Characters character);
 
     public void deleteCharacter(int characterId);
@@ -28,8 +28,7 @@ public interface CharactersDao{
     public List<Characters> getCharactersByOrganizationId(int organizationId);
 
     public List<Characters> getAllCharacters();
-    
+
     public List<Characters> getAssociatedCharacters(List<Sighting> sightings);
 
-    
 }
